+++ src/InitialWindow.ts
@@ -1,17 +1,8 @@
-import { UIManager } from 'react-native';
-import { Metrics } from './SafeArea.types';
+import { EdgeInsets, Metrics } from './SafeArea.types';
 
-const RNCSafeAreaViewConfig = UIManager.getViewManagerConfig(
-  'RNCSafeAreaView',
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-) as any;
-
-export const initialWindowMetrics = (RNCSafeAreaViewConfig != null &&
-RNCSafeAreaViewConfig.Constants != null
-  ? RNCSafeAreaViewConfig.Constants.initialWindowMetrics
-  : null) as Metrics | null;
+export const initialWindowMetrics: Metrics | null = null;
 
 /**
  * @deprecated
  */
-export const initialWindowSafeAreaInsets = initialWindowMetrics?.insets;
+export const initialWindowSafeAreaInsets: EdgeInsets | null = null;
