--- src/InitialWindow.web.ts
@@ -1,8 +1,0 @@
-import { EdgeInsets, Metrics } from './SafeArea.types';
-
-export const initialWindowMetrics: Metrics | null = null;
-
-/**
- * @deprecated
- */
-export const initialWindowSafeAreaInsets: EdgeInsets | null = null;
