+++ src/NativeSafeAreaView.native.tsx
@@ -1,0 +1,6 @@
+import { requireNativeComponent } from 'react-native';
+import { NativeSafeAreaViewProps } from './SafeArea.types';
+
+export default requireNativeComponent<NativeSafeAreaViewProps>(
+  'RNCSafeAreaView',
+);
